Move device fault diagnostics into backends
